Rename package to main
