+++ server/server.go
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package main
+package server
 
 import (
 	"fmt"
